Refactor PipelineNodes
